Fix files with a byte size larger than an int cause i/o errors (#3)
FIX: files with a byte size larger than an int cause i/o errors, closes #2
* Use long long for sizes, Use atoll(3) to read size values
* README file required by autotools

+++ support/grdiff.c
@@ -209,7 +209,7 @@ int read_stats(stats_t *stats, FILE *file){
 		}
 		if (gstrsub(line, "  Size") == 0){
 			// stats->size = atoi(line + strlen("  Size ");
-			stats->size = atoi(line + 7);
+			stats->size = atoll(line + 7);
 			size_set = 1;
 		};
 		if (strcmp(line, "  Type reg\n") == 0){
